Paper published at Thorax Journal_DOI:10.1136/thoraxjnl-2021-218194
A better understanding of the mechanism of action of mesenchymal stromal cells (MSCs) and their extracellular vesicles (EVs) is needed to support their use as novel therapies for acute respiratory distress syndrome (ARDS). Macrophages are important mediators of ARDS inflammatory response. Suppressor of cytokine signalling (SOCS) proteins are key regulators of the macrophage phenotype switch. We therefore investigated whether SOCS proteins are involved in mediation of the MSC effect on human macrophage reprogramming.