Abstract published at Cytotheraphy_DOI: https://doi.org/10.1016/S1465-3249(22)00145-1
MSC-derived extracellular vesicles (EVs) are considered as a cell-free therapy for ARDS. Previously we demon- strated that MSC-EVs contain functional mitochondria and are able to restore activity of the injured cells through mitochondrial transfer. However, the impact of this mechanism on the balance of oxidative phosphorylation and glycolysis is not known. We hypothesized that therapeutic potential of MSC-EVs will depend on their ability to alle- viate metabolic changes in primary human pulmonary cells exposed to LPS or ARDS plasma.